// core/types.ts
export interface HepatitisBData {
    agshb: 'si' | 'no';
    aghbc: 'si' | 'no';
}

export interface CoInfeccionData {
  tb: string;
  hepatitisB: HepatitisBData;
  hepatitisC: string;
  toxoplasmosis: string;
  criptococoxis: string;
  histoplasmosis: string;
  candida: string;
  neurosifilis: 'si' | 'no';
  paracoccidiomicosis: string;
  cmv: string;
  ebv: string;
}